% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{H_average}
\alias{H_average}
\title{Calculates average hydrophobicity}
\usage{
H_average(hydro_scale, prot_sequence)
}
\arguments{
\item{hydro_scale}{Hydrophobicity scale in a form of list (Wimley_White octanol scale used as default (current_scale))}

\item{prot_sequence}{Protein (peptide) sequence given as string (FASTA format with no description line, just the sequence)}
}
\value{
Sequence-based average hydrophobicity
}
\description{
Calculates average hydrophobicity of a given peptide sequence based on a hydrophobicity scale
}
