% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{H_moment_cross}
\alias{H_moment_cross}
\title{Draws a cross-section of a hydrophobicity moment map for a given turn angle and window size}
\usage{
H_moment_cross(hydro_scale, prot_sequence, window, turn_angle)
}
\arguments{
\item{hydro_scale}{Hydrophobicity scale in a form of list (Wimley_White octanol scale used as default (current_scale))}

\item{prot_sequence}{Protein (peptide) sequence given as string (FASTA format with no description line, just the sequence)}

\item{window}{Size of window (3, 5, 7, 9)}

\item{turn_angle}{Helix turn angle in radians}
}
\value{
A vector of hydrophobic moment values for a given sequence (has a length of: (number of amino acid - window size)+1 )
}
\description{
Calculates and draws a cross-section of a hydrophobicity moment map for a given turn angle and a window size (3,5,7,or 9).
}
