% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{H_moment_map}
\alias{H_moment_map}
\title{Draws a map of hydrophobic moment of a peptide for a given window size}
\usage{
H_moment_map(hydro_scale, prot_sequence, window)
}
\arguments{
\item{hydro_scale}{Hydrophobicity scale in a form of list (Wimley_White octanol scale used as default (current_scale))}

\item{prot_sequence}{Protein (peptide) sequence given as string (FASTA format with no description line, just the sequence). Must be shorter than 30 amino acids.}

\item{window}{Size of window (3, 5, 7, 9)}
}
\value{
A vector of hydrophobic moment values for a given sequence (has a length of: (number of amino acid - window size)+1 )
}
\description{
Calculates and draws hydrophobic moment map for a given peptide sequence for a given hydrophobicity scale, and window size. The results are presented graphically as a map for the following coordinates: amino acid number (middle of window) and helical turn angle (in degrees). More about hydrophobic moment maps can be found in the \href{http://onlinelibrary.wiley.com/doi/10.1016/j.febslet.2013.06.054/full}{article}.
}
