% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{H_moment_window}
\alias{H_moment_window}
\title{Calculates normalized hydrophobicity moment of a peptide for a given window size}
\usage{
H_moment_window(hydro_scale, prot_sequence, window, turn_angle)
}
\arguments{
\item{hydro_scale}{Hydrophobicity scale in a form of list (Wimley_White octanol scale used as default (current_scale))}

\item{prot_sequence}{Protein (peptide) sequence given as string (FASTA format with no description line, just the sequence)}

\item{window}{Size of window (3, 5, 7, 9)}

\item{turn_angle}{Helix turn angle in radians}
}
\value{
A vector of hydrophobic moment values for a given sequence (has a length of: (number of amino acid - window size)+1 )
}
\description{
Calculates normalized hydrophobicity moment for a given peptide sequence for a given hydrophobicity scale, helix trun angle and window size. For the given peptide sequence, hydrophobic moment is calculated for a window (of 3 or 5 or 7 or 9 amino acid) assuming helical structure with a given turn angle.
}
